requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For Parquet files
scipy>=1.10.0

# Machine Learning & NLP
scikit-learn>=1.3.0
# sentence-transformers>=2.2.0  # Heavy dependency - optional
# transformers>=4.30.0  # Heavy dependency - optional  
# torch>=2.0.0  # Heavy dependency - optional
# spacy>=3.6.0  # Windows incompatible - using textblob + vader instead
textblob>=0.17.0
vaderSentiment>=3.3.2
nltk>=3.8.0

# Financial Data & APIs
yfinance>=0.2.0
ccxt>=4.0.0  # Crypto exchanges
coinbase-advanced-py>=1.8.0  # Coinbase API for crypto data
polygon-api-client>=1.12.0
alpaca-trade-api>=3.0.0
finnhub-python>=2.4.0
newsapi-python>=0.2.6

# Web Scraping & Social Media
requests>=2.31.0
beautifulsoup4>=4.12.0
praw>=7.7.0  # Reddit API
tweepy>=4.14.0
playwright>=1.35.0

# Database & Storage
pymongo>=4.4.0
redis>=4.6.0
sqlalchemy>=2.0.0

# Technical Analysis
ta>=0.10.0
# talib>=0.4.25  # Windows incompatible - using ta instead
vectorbt>=0.25.0

# Monte Carlo & Risk
arch>=6.0.0  # GARCH models
cvxpy>=1.3.0  # Portfolio optimization
quantlib>=1.31

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
reportlab>=4.0.0  # PDF generation
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
tqdm>=4.65.0
loguru>=0.7.0
pydantic>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Windows-compatible alternatives
textblob>=0.17.0
vaderSentiment>=3.3.2
